pyproject.toml
[project]
name = "blog-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "langgraph",
    "langchain-openai",
    "langchain-core",
    "python-dotenv",
    "tavily",
    "loguru"
]
